Add null check before deleting backup directory

+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.admin.v1/src/main/java/org/wso2/carbon/apimgt/rest/api/admin/v1/utils/RestApiAdminUtils.java
@@ -243,7 +243,9 @@ public class RestApiAdminUtils {
             }
             zipInputStream.closeEntry();
             zipInputStream.close();
-            FileUtils.deleteDirectory(backupDirectory);
+            if (backupDirectory != null) {
+                FileUtils.deleteDirectory(backupDirectory);
+            }
         } catch (APIManagementException | IOException e) {
             revertTenantThemeImportChanges(tenantDomain, existingTenantTheme);
             throw new APIManagementException(e.getMessage(),
